boss2 drops a key opens up into ship in cave level backgrounds are retiled to show that. cave prefabs ordered correctly to display level as intended

+++ TwoPiece/Assets/Scripts/tileBackground.cs
@@ -5,8 +5,8 @@ using System.Collections.Generic;
 public class tileBackground : MonoBehaviour {
 
     [SerializeField] private GameObject tile;
-    [SerializeField] private int xStart = 5;
-    [SerializeField] private int yStart = 5;
+    [SerializeField] private float xStart = 5;
+    [SerializeField] private float yStart = 5;
     [SerializeField] private int xOffset = 5;
     [SerializeField] private int yOffset = 2;
     [SerializeField] private int xLength = 15;
